Refactored event handlers
