+++ src/components/_global/BalTextInput/composables/useInputEvents.ts
@@ -1,11 +1,20 @@
+import { ref } from 'vue';
+
 import { HtmlInputEvent } from '@/types';
 
 export default function useInputEvents(props, emit, validate) {
   /**
+   * STATE
+   */
+  const isActive = ref(false);
+  const isHover = ref(false);
+  /**
    * EVENTS
    */
+
   function onBlur(event: HtmlInputEvent) {
     emit('blur', event.target.value);
+    isActive.value = false;
     if (props.validateOn === 'blur') validate(event.target.value);
   }
 
@@ -14,8 +23,27 @@ export default function useInputEvents(props, emit, validate) {
       const overflowProtectedVal = overflowProtected(event.target.value);
       if (overflowProtectedVal) event.target.value = overflowProtectedVal;
     }
+    isActive.value = true;
     emit('input', event.target.value);
     emit('update:modelValue', event.target.value);
+  }
+
+  function onClick(event: HtmlInputEvent) {
+    isActive.value = true;
+    emit('click', event);
+  }
+
+  function onFocus(event: HtmlInputEvent) {
+    isActive.value = true;
+    emit('focus', event);
+  }
+  function onMouseOver(event: HtmlInputEvent) {
+    isHover.value = true;
+    emit('mouseOver', event);
+  }
+  function onMouseLeave(event: HtmlInputEvent) {
+    isHover.value = false;
+    emit('mouseLeave', event);
   }
 
   function onKeydown(event: HtmlInputEvent): void {
@@ -42,8 +70,14 @@ export default function useInputEvents(props, emit, validate) {
   }
 
   return {
+    isActive,
+    isHover,
     onBlur,
     onInput,
-    onKeydown
+    onKeydown,
+    onClick,
+    onFocus,
+    onMouseOver,
+    onMouseLeave
   };
 }
